🐛 Fix :  Main.jsx 리팩토링
